Rename test `failures` package to `failure`
