+++ protobuf-plugin/plugin/src/test/java/org/spine3/gradle/protobuf/failure/Given.java
@@ -18,13 +18,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.gradle.protobuf.failures;
+package org.spine3.gradle.protobuf.failure;
 
 import com.sun.javadoc.RootDoc;
 import org.gradle.tooling.GradleConnector;
 import org.gradle.tooling.ProjectConnection;
 import org.junit.rules.TemporaryFolder;
-import org.spine3.gradle.protobuf.failure.FailureJavadocGenerator;
 
 import java.io.IOException;
 import java.io.InputStream;
